Moved use cases mains to a directory.
