+++ src/main/java/screens/LoginScreen.java
@@ -3,9 +3,9 @@ package screens;
 import entities.Seller;
 import entities.User;
 import entities.UserType;
-import order_use_case.BuyerMain;
+import use_cases_mains.BuyerMain;
 import order_use_case.DoesNotExistException;
-import selling_use_case.SellerMain;
+import use_cases_mains.SellerMain;
 import user_login_use_case.LoginController;
 import user_login_use_case.LoginFailed;
 import user_login_use_case.LoginResponseModel;
