+++ src/main/java/use_cases_mains/BuyerMain.java
@@ -1,4 +1,4 @@
-package order_use_case;
+package use_cases_mains;
 
 import order_history_use_case.OrderHistoryInputBoundary;
 import order_history_use_case.OrderHistoryInteractor;
@@ -13,6 +13,6 @@ public class BuyerMain {
         OrderHistoryOutputBoundary orderHistoryOutputBoundary = new OrderHistoryPresenter();
         OrderHistoryInputBoundary orderHistoryInteractor = new OrderHistoryInteractor(username, email, orderHistoryOutputBoundary);
         BuyerDefaultView buyerDefaultView = new BuyerDefaultView(username, email, orderHistoryInteractor);
-       // TODO: fill this later once we have all the order classe created.
+       // TODO: fill this later once we have all the order classes created.
     }
 }
