+++ src/main/java/use_cases_mains/GetMenusMain.java
@@ -1,6 +1,7 @@
-package get_menus_use_case;
+package use_cases_mains;
 
 import entities.OrderFactory;
+import get_menus_use_case.*;
 import order_history_use_case.OrderHistoryInputBoundary;
 import order_history_use_case.OrderHistoryInteractor;
 import order_history_use_case.OrderHistoryOutputBoundary;
