+++ src/main/java/use_cases_mains/SellerMain.java
@@ -1,4 +1,4 @@
-package selling_use_case;
+package use_cases_mains;
 
 
 import order_use_case.DoesNotExistException;
@@ -6,6 +6,10 @@ import order_use_case.OrderDsGateway;
 import order_use_case.OrderGateway;
 import screens.SellerFulfillingOrderScreen;
 import screens.SellingScreen;
+import selling_use_case.SellingController;
+import selling_use_case.SellingInputBoundary;
+import selling_use_case.SellingInteractor;
+import selling_use_case.SellingPresenter;
 import user_access_use_case.SignUpDsGateway;
 import user_access_use_case.SignUpGateway;
